Move a type definition to a more central file
